refactor(daisi): Add new base manager class
