+++ daisi/src/natter-ns3/CMakeLists.txt
@@ -75,7 +75,7 @@ target_sources(daisi_natter_manager
 )
 target_link_libraries(daisi_natter_manager
         PUBLIC
-            daisi_manager
+            daisi_manager_old
             daisi_natter_application
             daisi_natter_static_network_calculation
             daisi_natter_scenariofile
